[cloud_firestore_web] Ensure Query pagination methods encode values.
This change ensures that the package:firestore Query objects are created
with properly encoded values when done so from the start/end
(pagination) methods.

This also adds some Timestamps to the unit tests to verify the behavior of
Timestamp -> DateTime.

There are some integration tests in cloud_firestore that cover this
behavior further.

Fixes https://github.com/FirebaseExtended/flutterfire/issues/2153