Update comment over name field in Firebase App
When configuring a Firebase App, if the name contains a ' ' (whitespace) instead of suggested separators, iOS throws a cryptic error message only visible through -v flag. Even then, the lack of context of the error makes it confusing to understand if the name causing the error was in the Info.plist or anywhere else. This additional comment would be really helpful.