fix(android,remote_config): ensureInitialized() task should ignore return value (fixes #5222)
Currently on Android the task returned by RemoteConfig.ensureInitialized() returns a `FirebaseRemoteConfigInfo` instance which gets sent to the Flutter standard message codec and causes a crash as this isn't a supported value that can be serialized by the codec. As we don't at this point care about the value of the task result, I've wrapped then task in a `whenAll` which will just return void when completed.