test (firebase_performance): complete e2e tests and fix some minor issues
Also update the perf readme with instructions on contributing to the code.

Currently the attributes related tests are failing because of a bug in the attribute methods of MethodChannelTrace class.

Somehow when running the e2e test on iOS there is also an "Failed to stop app" error, but the non-attributes related tests still passed.