fix(firebase_performance): fix the token verification.
This issue causes the example app to throw the error "'Platform interfaces must not be implemented with `implements`'" from PlatformInterface.verifyToken function call when running on the web browser. It will also cause the web plugin unit test (coming in the next PR) to fail due to the same error.