fix(ui): Web recaptcha hover removed after use.
Fixes #4377

After showing a recaptcha in web for SMS login, the recaptcha icon
overlay persists on the app until it is reloaded, often blocking an
app's buttons and other interactive widgets. This adds removal of the
web UI element in the clear method, and ensures that clear is called on
any RecaptchaVerifier() objects created by flutterfire.