Fix swizzle check for UNUserNotificationCenter
The shouldReplaceDelegate was originally added to prevent an infinite
loop due to double swizzling. This only happens however if the app
developer set UNUserNotificationCenterDelegate to the AppDelegate. This
check was added in PR #4043. However this check does not consider if the
non-nil was set to something else, either by the app developer or
another SDK. FluterFire should swizzle in this case, so it does not miss
these events.

To address the above this commit checks if notificationCenter.delegate
conformsToProtocol FlutterAppLifeCycleProvider. This forwards events so
this is the true protocol that should be checked to prevent the infinite
loop the PR #4043 was addressing.