fix(test): get tests to pass
The prefer_const_constructors warning is ignored in this case because the LoadingIndicator widget used inside the children list of the Row does not have a const constructor. This is due to the fact that LoadingIndicator extends PlatformWidget, which doesn't have a const constructor, making it impossible to have a const constructor for LoadingIndicator. Since we can't make the LoadingIndicator constructor const, we have to ignore the warning for the specific Row constructor. This is an acceptable solution in cases like this, where making the constructor const is not feasible or possible.