Allow serializing to Developer API models
Move parsing from top level methods, and serialization from a method on
the model,into instance methods on a new `SerializationStrategy` class.
Use the existing parsing code for the vertex strategy and copy over code
form the developer SDK for the Google AI strategy.

Add a `_ModelUri` class to allow a vertex and developer specialization
for the different backends.

Copy the serialization and parsing tests from the developer SDK, skip a
few tests that rely on arguments or methods that are unsupported in the
vertex model. Tests skipped for content embedding and extra arguments
passed to countTokens.