fix: a bug with the `demoProjectId` arg to `Firebase.initializeApp()`
If platform specific configuration files exist for the project (i.e.
GoogleService-Info.plist for iOS), the default `MethodChannelFirebaseApp` is
initialized using the configuration it specifies. This is fine since multiple
`FirebaseApp` instances can exist, however when
`MethodChannelFirebase.initializeApp()` is called it uses the default app name
(`[DEFAULT]`, from the code) if no name is passed as an argument. Since the
default app is initialized using the platform specific configuration before
`initializeApp()` is called, the user provided options are ignored.

The solution is to ensure a non-null app name is passed to `initializeApp()` to
avoid conflicts with platform specific configuration files when the user intends
to use a "demo-" project for testing.

Since we should be providing a distinct app name for the demo app, the
allowance for mismatched options between the specified options and the existing
options for demo apps is also removed.